Add files via upload
A simple security system program using tkinter as the GUI and a motion sensor. Its uses an interrupt triggered from a rising edge on a GPIO. I used GPIO 20/ pin 38 on Raspberry Pi b+ to do this, using the Raspbian os. Its programmed using Python 2 on the IDLE programming environment. Arming the system turns the trigger event on (sensing the rising edge) and disarming turns the event detection off.